fix : 장소 선택 default 로직 추가 & 리펙토링
