+++ src/components/Map/hooks/index.ts
@@ -1,4 +1,3 @@
 export { default as useMap } from './useMap';
-export { default as useMapListMarkers } from './useMapListMarkers';
+export { default as useSearchList } from './useSearchList';
 export { default as useMapSearchForm } from './useMapSearchForm';
-export { default as useMapPickMarker } from './useMapPickMarker';
